警告削除

+++ project/src/scene/play/chara/CharaBase.h
@@ -319,7 +319,7 @@ private:
 	Alarm*			m_Alarm;				// アラーム
 	int				m_hTrailImage;			// トレイルの画像ハンドル
 	int				m_Index;				// 自身のインデックス
-	int				m_HitPoint;
+	float			m_HitPoint;
 	float			m_BallChargeRate;		// ボールのチャージ加速度
 	float			m_MoveSpeed;			// 移動速度
 	float			m_RotSpeed;				// 回転速度
